courses/data_analysis/deepdive/bigtable-exercises/python/requirements.txt
google-cloud-bigtable==0.29.0
google-cloud-core==0.28.1
